requirements.txt
streamlit
requests
tenacity
openai
google-generativeai
